pyproject.toml
[project]
name = "kii"
version = "0.1.0"
description = "Python Script to Retrieve Data from phpMyAdmin Web Interface"
authors = [
    {name = "dmvaled",email = "dmvaled@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
